backend/my_django_app/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
